refactor csv reader includes because of r path length limitations:
