+++ src/execution/operator/csv_scanner/scanner/base_scanner.cpp
@@ -1,5 +1,5 @@
-#include "duckdb/execution/operator/csv_scanner/sniffer/csv_sniffer.hpp"
-#include "duckdb/execution/operator/csv_scanner/scanner/base_scanner.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_sniffer.hpp"
+#include "duckdb/execution/operator/csv_scanner/base_scanner.hpp"
 
 namespace duckdb {
