+++ src/execution/operator/csv_scanner/scanner/column_count_scanner.cpp
@@ -1,4 +1,4 @@
-#include "duckdb/execution/operator/csv_scanner/scanner/column_count_scanner.hpp"
+#include "duckdb/execution/operator/csv_scanner/column_count_scanner.hpp"
 
 namespace duckdb {
