+++ src/execution/operator/csv_scanner/scanner/string_value_scanner.cpp
@@ -1,7 +1,7 @@
-#include "duckdb/execution/operator/csv_scanner/scanner/string_value_scanner.hpp"
-#include "duckdb/execution/operator/csv_scanner/util/csv_casting.hpp"
-#include "duckdb/execution/operator/csv_scanner/scanner/skip_scanner.hpp"
-#include "duckdb/execution/operator/csv_scanner/table_function/csv_file_scanner.hpp"
+#include "duckdb/execution/operator/csv_scanner/string_value_scanner.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_casting.hpp"
+#include "duckdb/execution/operator/csv_scanner/skip_scanner.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_file_scanner.hpp"
 #include "duckdb/main/client_data.hpp"
 #include "duckdb/common/operator/integer_cast_operator.hpp"
 #include "duckdb/common/operator/double_cast_operator.hpp"
