+++ src/execution/operator/csv_scanner/sniffer/dialect_detection.cpp
@@ -1,4 +1,4 @@
-#include "duckdb/execution/operator/csv_scanner/sniffer/csv_sniffer.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_sniffer.hpp"
 #include "duckdb/main/client_data.hpp"
 
 namespace duckdb {
