+++ src/execution/operator/csv_scanner/sniffer/type_refinement.cpp
@@ -1,5 +1,5 @@
-#include "duckdb/execution/operator/csv_scanner/sniffer/csv_sniffer.hpp"
-#include "duckdb/execution/operator/csv_scanner/util/csv_casting.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_sniffer.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_casting.hpp"
 
 namespace duckdb {
