+++ src/execution/operator/csv_scanner/table_function/global_csv_state.cpp
@@ -1,7 +1,7 @@
-#include "duckdb/execution/operator/csv_scanner/table_function/global_csv_state.hpp"
+#include "duckdb/execution/operator/csv_scanner/global_csv_state.hpp"
 #include "duckdb/main/client_data.hpp"
-#include "duckdb/execution/operator/csv_scanner/scanner/scanner_boundary.hpp"
-#include "duckdb/execution/operator/csv_scanner/sniffer/csv_sniffer.hpp"
+#include "duckdb/execution/operator/csv_scanner/scanner_boundary.hpp"
+#include "duckdb/execution/operator/csv_scanner/csv_sniffer.hpp"
 #include "duckdb/execution/operator/persistent/csv_rejects_table.hpp"
 #include "duckdb/main/appender.hpp"
