+++ src/include/duckdb/storage/serialization/nodes.json
@@ -485,7 +485,7 @@
   {
     "class": "CSVOption<T>",
     "includes": [
-      "duckdb/execution/operator/csv_scanner/options/csv_option.hpp"
+      "duckdb/execution/operator/csv_scanner/csv_option.hpp"
     ],
     "members": [
       {
@@ -504,7 +504,7 @@
   {
     "class": "CSVReaderOptions",
     "includes": [
-      "duckdb/execution/operator/csv_scanner/options/csv_reader_options.hpp"
+      "duckdb/execution/operator/csv_scanner/csv_reader_options.hpp"
     ],
     "members": [
       {"id": 100,
